backend/requirements.txt
asgiref==3.2.10
bcrypt==3.2.0
certifi==2020.6.20
cffi==1.14.3
chardet==3.0.4
defusedxml==0.6.0
Django==3.1.1
django-allauth==0.42.0
django-cors-headers==3.5.0
djangorestframework==3.11.1
djangorestframework-jwt==1.11.0
idna==2.10
joblib==0.16.0
mysqlclient==2.0.1
numpy==1.19.2
oauthlib==3.1.0
pandas==1.1.2
pycparser==2.20
PyJWT==1.7.1
python-dateutil==2.8.1
python3-openid==3.2.0
pytz==2020.1
requests==2.24.0
requests-oauthlib==1.3.0
scikit-learn==0.23.2
scikit-surprise==1.1.1
scipy==1.5.2
simplejson==3.17.2
six==1.15.0
sklearn==0.0
sqlparse==0.3.1
surprise==0.1
threadpoolctl==2.1.0
urllib3==1.25.10
